refactor: organiza components
